branches/main: reverting to 699 revision
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@704 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ libs/include/cisstParameterTypes/prmString.h
@@ -55,7 +55,7 @@ protected:
 
 public:
     typedef prmString ThisType;
-    typedef size_t size_type;
+    typedef unsigned int size_type;
 
     /*! size of the generic Data payload,
      *  it should be big enough to carry your largest Data
