+++ libs/include/cisstVector/vctContainerTraits.h
@@ -30,8 +30,6 @@ http://www.cisst.org/cisst/license.txt.
 #ifndef _vctContainerTraits_h
 #define _vctContainerTraits_h
 
-#include <cstddef>
-
 /*! Macro used to define multiple types based on the type of elements.
   This will define size_type, index_type, difference_type,
   stride_type, value_type, reference, const_reference, pointer,
@@ -41,8 +39,8 @@ http://www.cisst.org/cisst/license.txt.
   \param type Type of element, e.g. double, float, char.
  */
 #define VCT_CONTAINER_TRAITS_TYPEDEFS(type) \
-    typedef size_t size_type; \
-    typedef size_t index_type; \
+    typedef unsigned int size_type; \
+    typedef unsigned int index_type; \
     typedef int difference_type; \
     typedef int stride_type; \
     typedef type value_type; \
